Consolidate cache.py into modeling_apriel2.py
Move all cache classes (_AttentionCache, _SSMCache, _DummyCacheLayer,
Apriel2Cache, _LayerListAccessor) into modeling_apriel2.py for better
tooling compatibility - modeling code is expected to be together.
