+++ fast_llm_external_models/tests/test_apriel2/test_modeling.py
@@ -62,7 +62,7 @@ class TestApriel2Modeling:
 
         # Test 1: Empty cache should give different results than filled cache
         # This verifies cache is being used at all
-        from fast_llm_external_models.apriel2.cache import Apriel2Cache, _AttentionCache
+        from fast_llm_external_models.apriel2.modeling_apriel2 import Apriel2Cache, _AttentionCache
 
         empty_cache = Apriel2Cache(config)
